+++ tests/Application/database/migrations/2021_09_10_000000_create_addresses_table.php
@@ -23,6 +23,8 @@ class CreateAddressesTable extends Migration
             $table->foreignIdFor(User::class, 'custom_foreign_id_for_name');
             $table->foreignIdFor(Address::class);
             $table->foreignIdFor(Address::class, 'nullable_address_id')->nullable();
+            $table->foreignId('foreign_id_constrained')->constrained('users');
+            $table->foreignId('nullable_foreign_id_constrained')->nullable()->constrained('users');
         });
     }
 }